+++ .devcontainer/dev/setup-nfs.sh
@@ -27,9 +27,9 @@ apt-get install -y -qq \
 	libreadline-dev
 
 cd /tmp
-curl -o nfs-utils-2.8.3.tar.gz https://mirrors.edge.kernel.org/pub/linux/utils/nfs-utils/2.8.3/nfs-utils-2.8.3.tar.gz
-tar xvf nfs-utils-2.8.3.tar.gz
-cd /tmp/nfs-utils-2.8.3
+curl -o nfs-utils-2.8.4.tar.gz https://mirrors.edge.kernel.org/pub/linux/utils/nfs-utils/2.8.4/nfs-utils-2.8.4.tar.gz
+tar xvf nfs-utils-2.8.4.tar.gz
+cd /tmp/nfs-utils-2.8.4
 
 ./configure \
 	--build=x86_64-linux-gnu \