+++ image-name.sh
@@ -4,7 +4,8 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-# extract the name of the AWS AMI that was built by Packer
+# extract the name of the AWS AMI(s) that were built by Packer
+# supports both amd64 and arm64 architectures
 exec jq -r '.builds[] |
-	select(.name == "nfs-proxy" and .builder_type == "amazon-ebs") |
-	.artifact_id' image.manifest.json
+	select((.name == "nfs-proxy-amd64" or .name == "nfs-proxy-arm64") and .builder_type == "amazon-ebs") |
+	(.name | sub("nfs-proxy-"; "")) + ":" + (.artifact_id | split(":")[1])' image.manifest.json